requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
SQLAlchemy>=2.0.29
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pendulum>=3.0.0
python_dateutil>=2.9.0
pytz>=2024.1
httpx>=0.27.0
tenacity>=9.0.0
cachetools>=5.3.0
google_api_python_client>=2.167.0
fastapi-mcp>=0.4.0
